requirements.txt
boto3==1.28.0
duckdb==0.5.1
matplotlib==3.8.2
memory_profiler==0.61.0
numpy==1.26.4
pandas==1.3.4
psycopg2-binary==2.9.6
Requests==2.31.0
SQLAlchemy==1.4.22
pipreqs==0.4.13
cchardet~=2.1.7
openai~=1.25.0
python-dotenv~=1.0.1
torch~=1.13.1
transformers~=4.36.1
scikit-learn~=1.4.1.post1
tqdm~=4.66.1
seaborn~=0.13.0
scipy~=1.12.0
joblib~=1.3.2
Pillow~=10.0.0
astmonkey~=0.3.6